fix error + increase readability
The main script contained an error in the multiplication in section 5. The sum should have *not* added '1101' shifted left by two. This change increases the correlation between multiplication and the sum of the first operand shifted left *in accordance with* the second. If there is a zero bit in the second then you need to multiply by 0